requirements.txt
numpy==1.26.4
pandas==2.2.3
scikit-learn==1.5.2
PyYAML==6.0.2
torch==2.10.0
fastapi==0.116.1
uvicorn==0.35.0
pydantic==2.10.6
plotly==5.24.1
pytest==8.3.4
